requirements.txt
discord.py
aiohttp
python-dotenv
notion-client